Fix labs warning

+++ lwadvapi/threaded/lwtime.c
@@ -227,7 +227,7 @@ LwSetSystemTime(
     }
         
     //Make sure the time is now within 5 seconds of what we set
-    if (labs(readTime - ttCurTime) > 5)
+    if (labs((long)(readTime - ttCurTime)) > 5)
     {
         LW_RTL_LOG_ERROR("Attempted to set time to %ld, but it is now %lld.", ttCurTime, readTime);
         dwError = LW_ERROR_FAILED_TO_SET_TIME;
@@ -240,7 +240,7 @@ LwSetSystemTime(
     {
         readTime = time(NULL);
 
-        if (labs(readTime - ttCurTime) > 5)
+        if (labs((long)(readTime - ttCurTime)) > 5)
         {
             LW_RTL_LOG_DEBUG("Time is slow to update...waiting");
             sleep(1);
